rename the package 'syngen/ml/pipeline' to 'syngen/ml/utils'
