+++ src/syngen/ml/reporters/reporters.py
@@ -5,7 +5,7 @@ import pandas as pd
 import numpy as np
 from loguru import logger
 
-from syngen.ml.pipeline import (
+from syngen.ml.utils import (
     get_nan_labels,
     nan_labels_to_float,
     fetch_dataset
